src/requirements.txt
pyserial
matplotlib
redis
